moved files to action page
